remove record deletes that serve no purpose
since the fundeps are set up the correct way in general, there's no need to
actaully pop fields off the records as we iterate them. so we take advantage of
the fact that the row types are not used for resolution to throw them away, only
caring about the few row types that we use to build records.